requirements.txt
asn1crypto==1.2.0
certifi==2019.9.11
cffi==1.13.1
cryptography==2.8
Django==2.1.4
django-cors-headers==3.1.1
django-rest-multiple-models==2.1.3
djangorestframework==3.10.3
idna==2.8
mysqlclient==1.4.4
Pillow==6.2.1
pycparser==2.19
PyMySQL==0.9.3
pytz==2019.3
six==1.12.0
sqlparse==0.3.0
wincertstore==0.2
